Refactor namespace to PSR
