+++ tests/Unit/Client/MultiClientTest.php
