+++ tests/Unit/Client/StatusWatcherClient/AnonymousMessageMock.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Tests\Client\StatusWatcherClient;
+namespace Unit\Client\StatusWatcherClient;
 
 use TelegramOSINT\Exception\TGException;
 use TelegramOSINT\MTSerialization\AnonymousMessage;
