+++ tests/Unit/Factor/FactorizationTest.php
