+++ tests/Unit/MTSerialization/OwnImplementation/expected_objects/foreign_user
