+++ tests/Unit/MTSerialization/OwnImplementation/expected_objects/get_config
