+++ tests/Unit/MTSerialization/OwnImplementation/expected_objects/gzipped_content
