+++ tests/Unit/MTSerialization/OwnImplementation/expected_objects/rpc_result_with_object_response
