+++ tests/Unit/MTSerialization/OwnImplementation/expected_objects/tg_app/get_config
