+++ tests/Unit/MTSerialization/OwnImplementation/expected_objects/tg_app/schema_json_optional_arguments_using_flags
