+++ tests/Unit/MTSerialization/OwnImplementation/expected_objects/tg_app/update_contacts_reset
