+++ tests/Unit/MTSerialization/OwnImplementation/expected_objects/tg_app/update_link_response
