+++ tests/Unit/MTSerialization/OwnImplementation/expected_objects/tg_app/update_new_channel_message
