+++ tests/Unit/MTSerialization/OwnImplementation/expected_objects/update_on_new_photo
