+++ tests/Unit/MTSerialization/OwnImplementation/expected_objects/usual_msg_container_deserialization
