+++ tests/Unit/Registration/AccountInfoTest.php
