+++ tests/Unit/TLMessage/TLMessage/ClientMessages/TgAppSerializationTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Tests\TLMessage\TLMessage\ClientMessages;
+namespace Unit\TLMessage\TLMessage\ClientMessages;
 
 use PHPUnit\Framework\TestCase;
 use TelegramOSINT\TLMessage\TLMessage\ClientMessages\Shared\get_config;
