+++ tests/phpunit.config.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8" ?>
-<phpunit bootstrap="TestClassLoader.php">
+<phpunit bootstrap="../vendor/autoload.php">
   <filter>
     <whitelist processUncoveredFilesFromWhitelist="true">
       <directory suffix=".php">../src</directory>
